add validation & picture store

+++ routes/web.php
@@ -17,10 +17,10 @@ Auth::routes();
 
 Route::get('/', 'HomeController@index');
 Route::get('/home', 'HomeController@index')->name('home');
-Route::get('/profile', 'UserController@profile');
-Route::put('/uploadPhoto', 'UserController@store');
-Route::post('/updateProfile', 'UserController@update');
-Route::post('/changePass', 'UserController@changePass');
+Route::get('/profile', 'UserController@profile')->middleware('auth');
+Route::put('/uploadPhoto', 'UserController@store')->middleware('auth');
+Route::post('/updateProfile', 'UserController@update')->middleware('auth');
+Route::post('/changePass', 'UserController@changePass')->middleware('auth');
 Route::get('/resetPassword', 'UserController@vResetPass');
 Route::post('/resetPassword', 'UserController@resetPass');
 
@@ -28,41 +28,41 @@ Route::get('/search', 'ProductController@search');
 Route::get('/product/detail/{productId}', 'ProductController@show');
 Route::get('/product/type/{typeId}', 'ProductController@type');
 Route::get('/product/category/{categoryId}', 'ProductController@category');
-Route::post('/addProduct', 'ProductController@store');
-Route::post('/updateProduct/{productId}', 'ProductController@update');
-Route::delete('/deleteProduct/{productId}', 'ProductController@destroy');
+Route::post('/addProduct', 'ProductController@store')->middleware('auth', 'isVerif', 'isSeller');
+Route::post('/updateProduct/{productId}', 'ProductController@update')->middleware('auth', 'isVerif', 'isSeller');
+Route::delete('/deleteProduct/{productId}', 'ProductController@destroy')->middleware('auth', 'isVerif', 'isSeller');
 
-Route::post('/create/store', 'StoreController@create');
-Route::get('/store/info/{storeId}', 'StoreController@show');
-Route::put('/edit/store/{storeId}', 'StoreController@edit');
-Route::get('/order/{storeId}', 'StoreController@order');
-Route::put('/done/{transactionId}', 'TransactionController@done');
-Route::put('/cancel/{transactionId}', 'TransactionController@cancel');
+Route::post('/create/store', 'StoreController@create')->middleware('auth', 'isVerif');
+Route::get('/store/info/{storeId}', 'StoreController@show')->middleware('auth', 'isVerif', 'isSeller');
+Route::put('/edit/store/{storeId}', 'StoreController@edit')->middleware('auth', 'isVerif', 'isSeller');
+Route::get('/order/{storeId}', 'StoreController@order')->middleware('auth', 'isVerif', 'isSeller');
+Route::put('/done/{transactionId}', 'TransactionController@done')->middleware('auth', 'isVerif', 'isSeller');
+Route::put('/cancel/{transactionId}', 'TransactionController@cancel')->middleware('auth', 'isVerif', 'isSeller');
 
-Route::post('/send/review/{productId}', 'ReviewController@review');
+Route::post('/send/review/{productId}', 'ReviewController@review')->middleware('auth');
 
-Route::get('/cart/{userId}', 'CartController@show');
-Route::post('/addToCart/{productId}/{userId}', 'CartController@store');
-Route::delete('/delete/cart', 'CartController@destroy');
-Route::post('/checkout', 'CartController@checkout');
+Route::get('/cart/{userId}', 'CartController@show')->middleware('auth', 'isBuyer');
+Route::post('/addToCart/{productId}/{userId}', 'CartController@store')->middleware('auth', 'isBuyer');
+Route::delete('/delete/cart', 'CartController@destroy')->middleware('auth', 'isBuyer');
+Route::post('/checkout', 'CartController@checkout')->middleware('auth', 'isBuyer');
 
-Route::post('/buy/{productId}', 'ProductController@cBuy');
-Route::get('/buy/{productId}', 'ProductController@vBuy');
-Route::put('/buy/{transactionId}', 'ProductController@buy');
+Route::post('/buy/{productId}', 'ProductController@cBuy')->middleware('auth', 'isBuyer');
+Route::get('/buy/{productId}', 'ProductController@vBuy')->middleware('auth', 'isBuyer');
+Route::put('/buy/{transactionId}', 'ProductController@buy')->middleware('auth', 'isBuyer');
 
-Route::get('/transber', 'TransberController@index');
-Route::get('/transber/detail/{transberId}', 'TransberController@show');
+Route::get('/transber', 'TransberController@index')->middleware('auth', 'isBuyer');
+Route::get('/transber/detail/{transberId}', 'TransberController@show')->middleware('auth', 'isBuyer');
 
-Route::get('/transaction', 'TransactionController@index');
-Route::get('/transaction/history/{userId}', 'TransactionController@show');
-Route::put('/doneTransaction/{transactionId}', 'TransactionController@doneTransaction');
-Route::delete('/cancelTransaction/{transactionId}', 'TransactionController@cancelTransaction');
-Route::delete('/deleteTransaction/{transactionId}', 'TransactionController@destroy');
+Route::get('/transaction', 'TransactionController@index')->middleware('auth');
+Route::get('/transaction/history/{userId}', 'TransactionController@show')->middleware('auth');
+Route::put('/doneTransaction/{transactionId}', 'TransactionController@doneTransaction')->middleware('auth');
+Route::delete('/cancelTransaction/{transactionId}', 'TransactionController@cancelTransaction')->middleware('auth');
+Route::delete('/deleteTransaction/{transactionId}', 'TransactionController@destroy')->middleware('auth');
 
-Route::get('/verifKTP', 'UserController@vVerifKTP');
-Route::post('/veritKTP', 'UserController@sVerifKTP');
+Route::get('/verifKTP', 'UserController@vVerifKTP')->middleware('auth');
+Route::post('/veritKTP', 'UserController@sVerifKTP')->middleware('auth');
 
-Route::post('/transber/{category}', 'TransberController@transber');
-Route::put('/transber/payment/{transberId}', 'TransberController@payment');
-Route::delete('/cancelTransber/{transberId}', 'TransberController@destroy');
-Route::put('/done/{roleTransber}/{transberId}', 'TransberController@done');
+Route::post('/transber/{category}', 'TransberController@transber')->middleware('auth', 'isBuyer');
+Route::put('/transber/payment/{transberId}', 'TransberController@payment')->middleware('auth', 'isBuyer');
+Route::delete('/cancelTransber/{transberId}', 'TransberController@destroy')->middleware('auth', 'isBuyer');
+Route::put('/done/{roleTransber}/{transberId}', 'TransberController@done')->middleware('auth', 'isBuyer');
